feat(TournamentsListing): add byYear handling to commons
